Чиню сигнатуру protectFd: gomobile перевёл Go int в Long, не в Int
Реальная сборка libXray.aar показала точную сигнатуру интерфейса:
fun protectFd(p0: Long): Boolean — gomobile мапит Go int в Kotlin/Java
long (безопаснее, раз в Go int может быть 32 или 64 бита), я в
XrayCoreBridge.kt предполагал Int. protect() в Android API как был Int,
так и остался — добавил toInt() на границе.

+++ app/src/main/kotlin/com/stailegrow/maxstrike/vpn/MaxStrikeVpnService.kt
@@ -57,7 +57,9 @@ class MaxStrikeVpnService : VpnService(), libXray.DialerController {
     private var tunInterface: ParcelFileDescriptor? = null
     private var coreRunning = false
 
-    override fun protectFd(fd: Int): Boolean = protect(fd)
+    // gomobile переводит Go-тип int в Kotlin как Long (не Int) — сам
+    // адрес protect() в Android-API остаётся Int, поэтому toInt() ниже.
+    override fun protectFd(fd: Long): Boolean = protect(fd.toInt())
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         when (intent?.action) {
